Fix #1639 . Fix issue with too many wings. Change all remaining Item equip slot fields from sbyte to int.

+++ ExampleMod/Content/Items/Accessories/WaspNest.cs
@@ -44,7 +44,7 @@ namespace ExampleMod.Content.Items.Accessories
 		}
 
 		public override void SetDefaults() {
-			sbyte realBackSlot = Item.backSlot;
+			int realBackSlot = Item.backSlot;
 			Item.CloneDefaults(ItemID.HiveBackpack);
 			Item.value = Item.sellPrice(0, 5);
 			// CloneDefaults will clear out the autoloaded Back slot, so we need to preserve it this way.
